+++ delog-examples/examples/try-log.rs
@@ -1,10 +1,13 @@
 #[macro_use]
 extern crate delog;
 
-use delog::flushers::StdoutFlusher;
-use delog::renderers::RipgrepRenderer;
+use delog::{
+    example::StdoutFlusher,
+    render::RipgrepRenderer,
+};
+
 delog!(Delogger, 196, StdoutFlusher, renderer: RipgrepRenderer);
-local_macros!();
+generate_macros!();
 
 static STDOUT_FLUSHER: StdoutFlusher = StdoutFlusher {};
 static RENDERER: RipgrepRenderer = RipgrepRenderer {};
@@ -12,22 +15,23 @@ static RENDERER: RipgrepRenderer = RipgrepRenderer {};
 fn main() {
     Delogger::init(delog::LevelFilter::Info, &STDOUT_FLUSHER, &RENDERER).ok();
 
-    // do some serious work
-    global_try_warn!("This is a warning").unwrap();
-    global_try_info!("This is information").unwrap();
-    global_try_warn!("This is a warning").unwrap();
-    global_try_info!("This is information").expect_err("should error out due to incapacity");
+    // // do some serious work
+    // global_try_warn!("This is a warning").unwrap();
+    // global_try_info!("This is information").unwrap();
+    // global_try_warn!("This is a warning").unwrap();
+    // global_try_info!("This is information").expect_err("should error out due to incapacity");
 
-    // flush the logs
-    Delogger::flush();
+    // // flush the logs
+    // Delogger::flush();
 
-    println!("===");
+    // println!("===");
 
     try_warn!("This is a warning").unwrap();
+    try_info!(target: "!", "This is information NOW").unwrap();
     try_info!("This is information").unwrap();
-    try_warn!(target: "!", "This is a warning").unwrap();
+    try_warn_now!("This is a warning").unwrap();
     try_warn!("This is a warning").unwrap();
-    #[cfg(not(feature = "log-none"))]
+    #[cfg(not(any(feature = "log-none", feature = "log-warn")))]
     try_info!("This is information").expect_err("should error out due to incapacity");
     #[cfg(feature = "log-none")]
     try_info!("This is information").ok();