+++ gate-tests/lib-a/lib-a1/src/lib.rs
@@ -1,7 +1,21 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
+delog::generate_macros!();
+
+pub fn f() {
+    submodule::sub_f()
+}
+
+pub mod submodule {
+    pub fn sub_f() {
+        log::info!("global info log from lib_a1");
+        log::warn!("global info log from lib_a1");
+
+        // log!(delog::Level::Info, "log level info from lib_a::f");
+        // log!(target: "!", delog::Level::Info, "log level info from lib_a::f");
+        // log!(target: "!", delog::Level::Warn, "log level warn from lib_a::f");
+
+        info!("local info from lib_a1::f");
+        warn!("local warn from lib_a1::f");
+        info_now!("immediate local info from lib_a1::f");
+        warn_now!("immediate local warn from lib_a1::f");
     }
 }