+++ src/renderers.rs
@@ -5,23 +5,23 @@ use crate::Renderer;
 
 #[derive(Clone, Copy)]
 /// Renders just the `record.args()`.
-pub struct ArgumentsRenderer {}
+pub struct DefaultRenderer {}
 
 /// The default, minimal renderer.
-pub fn default() -> &'static ArgumentsRenderer {
-    static RENDERER: ArgumentsRenderer = ArgumentsRenderer {};
+pub fn default() -> &'static DefaultRenderer {
+    static RENDERER: DefaultRenderer = DefaultRenderer {};
     &RENDERER
 }
 
 
-impl Renderer for ArgumentsRenderer {
+impl Renderer for DefaultRenderer {
     fn render<'a>(&self, buf: &'a mut [u8], record: &log::Record) -> &'a [u8] {
         render_arguments(buf, *record.args())
     }
 }
 
-unsafe impl Send for ArgumentsRenderer {}
-unsafe impl Sync for ArgumentsRenderer {}
+unsafe impl Send for DefaultRenderer {}
+unsafe impl Sync for DefaultRenderer {}
 
 #[derive(Clone, Copy)]
 /// Renders the `record.args()`, prefixed by level, target, and file, line if they are some.